Link separate JSystem libraries due to control.cpp
This allows the Xcode generator to work, otherwise
it breaks on the duplicate control.o files within
the game_debug library.
