+++ src/dusk/extras.c
@@ -3,6 +3,27 @@
 #include <string.h>
 #include <stdint.h>
 
+#ifdef _MSC_VER
+#include <intrin.h>
+#endif
+
+void __dcbz(void* addr, int offset) {
+    // Gekko cache lines are 32 bytes.
+    // dcbz usually requires addr to be 32-byte aligned.
+    memset((char*)addr + offset, 0, 32);
+}
+
+int __cntlzw(unsigned int val) {
+    if (val == 0) return 32; // PowerPC returns 32 if the input is 0
+#ifdef _MSC_VER
+    unsigned long idx;
+    _BitScanReverse(&idx, val);
+    return 31 - (int)idx;
+#else
+    return __builtin_clz(val);
+#endif
+}
+
 #ifndef _MSC_VER
 int stricmp(const char* str1, const char* str2) {
 	char a_var;
@@ -47,11 +68,6 @@ int strnicmp(const char* str1, const char* str2, int n) {
     return 0;
 }
 #endif
-
-
-void *_memcpy(void* dest, void const* src, int n) {
-    return memcpy(dest, src, n);
-}
 
 void DCZeroRange(void* addr, uint32_t nBytes) {
 #if defined(_MSC_VER) || TARGET_ANDROID
